packages/common/src/experiences/guest-experience/state/cms-content/cms-content-wrappers/plan-member-support/plan-member-support.cms-content-wrapper.test.ts
// Copyright 2022 Prescryptive Health, Inc.

import { IPlanMemberSupportContent } from '../../../../../../models/cms-content/plan-member-support.content';
import { defaultLanguage } from '../../../../../../models/language';
import {
  IUIContent,
  IUIContentGroup,
} from '../../../../../../models/ui-content';
import {
  findContentValue,
  getContent,
} from '../../../../../../utils/content/cms-content-wrapper.helper';
import { CmsGroupKey } from '../../cms-group-key';
import { planMemberSupportCMSContentWrapper } from './plan-member-support.cms-content-wrapper';

jest.mock('../../../../../../utils/content/cms-content-wrapper.helper');
const findContentValueMock = findContentValue as jest.Mock;
const getContentMock = getContent as jest.Mock;

describe('planMemberSupportCMSContentWrapper', () => {
  it('has correct content', () => {
    const expectedCoBrandingCMSContentMock: IPlanMemberSupportContent = {
      title: 'title',
    };

    findContentValueMock.mockReturnValueOnce(
      expectedCoBrandingCMSContentMock.title
    );

    const uiContentMock: IUIContent[] = [
      {
        fieldKey: 'field-key',
        language: 'English',
        type: 'Text',
        value: 'value',
      },
    ];
    getContentMock.mockReturnValue(uiContentMock);

    const cmsContentMapMock: Map<string, IUIContentGroup> = new Map([
      [
        CmsGroupKey.planMemberSupport,
        {
          content: uiContentMock,
          lastUpdated: 0,
          isContentLoading: true,
        },
      ],
    ]);

    const result = planMemberSupportCMSContentWrapper(
      defaultLanguage,
      cmsContentMapMock
    );

    expect(getContentMock).toHaveBeenCalledWith(
      defaultLanguage,
      cmsContentMapMock,
      CmsGroupKey.planMemberSupport
    );

    const expectedFieldKeys = ['title'];
    expect(findContentValueMock).toHaveBeenCalledTimes(
      expectedFieldKeys.length
    );
    expectedFieldKeys.forEach((key, index) => {
      expect(findContentValueMock).toHaveBeenNthCalledWith(
        index + 1,
        key,
        uiContentMock
      );
    });

    expect(result).toEqual(expectedCoBrandingCMSContentMock);
  });
});
